+++ util/src/main/java/com/google/cloud/hadoop/util/AbstractGoogleAsyncWriteChannel.java
@@ -19,7 +19,6 @@ import com.google.api.client.http.InputStreamContent;
 import com.google.common.annotations.VisibleForTesting;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.PipedInputStream;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
 
@@ -50,7 +49,7 @@ public abstract class AbstractGoogleAsyncWriteChannel<T extends AbstractGoogleCl
   public abstract T createRequest(InputStreamContent inputStream) throws IOException;
 
   @Override
-  public void startUpload(PipedInputStream pipeSource) throws IOException {
+  public void startUpload(InputStream pipeSource) throws IOException {
     // Connect pipe-source to the stream used by uploader.
     InputStreamContent objectContentStream = new InputStreamContent(contentType, pipeSource);
     // Indicate that we do not know length of file in advance.
@@ -61,7 +60,7 @@ public abstract class AbstractGoogleAsyncWriteChannel<T extends AbstractGoogleCl
     request.setDisableGZipContent(true);
 
     // Change chunk size from default value (10MB) to one that yields higher performance.
-    clientRequestHelper.setChunkSize(request, uploadChunkSize);
+    clientRequestHelper.setChunkSize(request, options.getUploadChunkSize());
 
     // Given that the two ends of the pipe must operate asynchronous relative
     // to each other, we need to start the upload operation on a separate thread.
@@ -85,7 +84,7 @@ public abstract class AbstractGoogleAsyncWriteChannel<T extends AbstractGoogleCl
     public S call() throws Exception {
       // Try-with-resource will close this end of the pipe so that
       // the writer at the other end will not hang indefinitely.
-      try (InputStream uploadPipeSource = pipeSource) {
+      try (InputStream ignore = pipeSource) {
         return uploadObject.execute();
       } catch (IOException ioe) {
         S response = createResponseFromException(ioe);