+++ util/src/main/java/com/google/cloud/hadoop/util/BaseAbstractGoogleAsyncWriteChannel.java
@@ -17,14 +17,16 @@ package com.google.cloud.hadoop.util;
 import static com.google.common.base.Preconditions.checkState;
 
 import com.google.common.flogger.GoogleLogger;
+import java.io.BufferedInputStream;
 import java.io.IOException;
-import java.io.OutputStream;
+import java.io.InputStream;
 import java.io.PipedInputStream;
 import java.io.PipedOutputStream;
 import java.nio.ByteBuffer;
 import java.nio.channels.Channels;
 import java.nio.channels.ClosedByInterruptException;
 import java.nio.channels.ClosedChannelException;
+import java.nio.channels.Pipe;
 import java.nio.channels.WritableByteChannel;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
@@ -47,22 +49,14 @@ public abstract class BaseAbstractGoogleAsyncWriteChannel<T> implements Writable
   // write channel (pipeSinkChannel below). The pipe is formed by connecting pipeSink to pipeSource
   protected final ExecutorService threadPool;
 
-  private boolean initialized = false;
-
-  // Size of buffer used by upload pipe.
-  private final int pipeBufferSize;
-
-  // Chunk size to use.
-  protected final int uploadChunkSize;
-
-  // A channel wrapper over pipeSink.
-  private WritableByteChannel pipeSinkChannel;
+  protected final AsyncWriteChannelOptions options;
 
   // Upload operation that takes place on a separate thread.
   protected Future<T> uploadOperation;
 
-  // When enabled, we get higher throughput for writing small files.
-  private final boolean directUploadEnabled;
+  private boolean initialized = false;
+
+  private WritableByteChannel pipeSink;
 
   private ByteBuffer uploadCache = null;
 
@@ -70,12 +64,10 @@ public abstract class BaseAbstractGoogleAsyncWriteChannel<T> implements Writable
   public BaseAbstractGoogleAsyncWriteChannel(
       ExecutorService threadPool, AsyncWriteChannelOptions options) {
     this.threadPool = threadPool;
-    this.pipeBufferSize = options.getPipeBufferSize();
+    this.options = options;
     if (options.getUploadCacheSize() > 0) {
       this.uploadCache = ByteBuffer.allocate(options.getUploadCacheSize());
     }
-    this.uploadChunkSize = options.getUploadChunkSize();
-    this.directUploadEnabled = options.isDirectUploadEnabled();
     setContentType("application/octet-stream");
   }
 
@@ -104,7 +96,7 @@ public abstract class BaseAbstractGoogleAsyncWriteChannel<T> implements Writable
 
   /** Returns true if direct media uploads are enabled. */
   public boolean isDirectUploadEnabled() {
-    return directUploadEnabled;
+    return options.isDirectUploadEnabled();
   }
 
   /**
@@ -141,7 +133,7 @@ public abstract class BaseAbstractGoogleAsyncWriteChannel<T> implements Writable
     }
 
     try {
-      return pipeSinkChannel.write(buffer);
+      return pipeSink.write(buffer);
     } catch (IOException e) {
       throw new IOException(
           String.format(
@@ -157,7 +149,7 @@ public abstract class BaseAbstractGoogleAsyncWriteChannel<T> implements Writable
    */
   @Override
   public boolean isOpen() {
-    return (pipeSinkChannel != null) && pipeSinkChannel.isOpen();
+    return (pipeSink != null) && pipeSink.isOpen();
   }
 
   /**
@@ -175,7 +167,7 @@ public abstract class BaseAbstractGoogleAsyncWriteChannel<T> implements Writable
       return;
     }
     try {
-      pipeSinkChannel.close();
+      pipeSink.close();
       handleResponse(waitForCompletionAndThrowIfUploadFailed());
     } catch (IOException e) {
       if (uploadCache == null) {
@@ -208,7 +200,7 @@ public abstract class BaseAbstractGoogleAsyncWriteChannel<T> implements Writable
   }
 
   private void closeInternal() {
-    pipeSinkChannel = null;
+    pipeSink = null;
     if (uploadOperation != null && !uploadOperation.isDone()) {
       uploadOperation.cancel(/* mayInterruptIfRunning= */ true);
     }
@@ -217,19 +209,33 @@ public abstract class BaseAbstractGoogleAsyncWriteChannel<T> implements Writable
 
   /** Initialize this channel object for writing. */
   public void initialize() throws IOException {
-    // Create a pipe such that its one end is connected to the input stream used by
-    // the uploader and the other end is the write channel used by the caller.
-    PipedInputStream pipeSource = new PipedInputStream(pipeBufferSize);
-    OutputStream pipeSink = new PipedOutputStream(pipeSource);
-    pipeSinkChannel = Channels.newChannel(pipeSink);
-
+    InputStream pipeSource = initializeUploadPipe();
     startUpload(pipeSource);
-
     initialized = true;
   }
 
+  // Create a pipe such that its one end is connected to the input stream used by
+  // the uploader and the other end is the write channel used by the caller.
+  private InputStream initializeUploadPipe() throws IOException {
+    switch (options.getPipeType()) {
+      case NIO_CHANNEL_PIPE:
+        Pipe pipe = Pipe.open();
+        pipeSink = pipe.sink();
+        InputStream pipeSource = Channels.newInputStream(pipe.source());
+        return options.getPipeBufferSize() > 0
+            ? new BufferedInputStream(pipeSource, options.getPipeBufferSize())
+            : pipeSource;
+      case IO_STREAM_PIPE:
+        PipedInputStream internalPipeSource = new PipedInputStream(options.getPipeBufferSize());
+        PipedOutputStream internalPipeSink = new PipedOutputStream(internalPipeSource);
+        pipeSink = Channels.newChannel(internalPipeSink);
+        return internalPipeSource;
+    }
+    throw new IllegalStateException("Unknown PipeType: " + options.getPipeType());
+  }
+
   /** Create a new thread which handles the upload. */
-  public abstract void startUpload(PipedInputStream pipeSource) throws IOException;
+  public abstract void startUpload(InputStream pipeSource) throws IOException;
 
   /** Sets the contentType. This must be called before {@link #initialize()} for any effect. */
   protected void setContentType(String contentType) {