+++ gcs/CHANGES.md
@@ -40,6 +40,22 @@
     wait between syncs. `hsync()` when rate limited will block on waiting for
     the permits, but `hflush()` will simply perform nothing and return.
 
+1.  Added a new parameter to configure output stream pipe type:
+
+    ```
+    fs.gs.outputstream.pipe.type (default: IO_STREAM_PIPE)
+    ```
+
+    Valid values are `NIO_CHANNEL_PIPE` and `IO_STREAM_PIPE`.
+
+    Output stream now supports (when property value set to `NIO_CHANNEL_PIPE`)
+    [Java NIO Pipe](https://docs.oracle.com/javase/8/docs/api/java/nio/channels/Pipe.html)
+    that allows to reliably write in the output stream from multiple threads
+    without *"Pipe broken"* exceptions.
+
+    Note that when using `NIO_CHANNEL_PIPE` option maximum upload throughput can
+    decrease by 10%.
+
 ### 2.1.1 - 2020-03-11
 
 1.  Add upload cache to support high-level retries of failed uploads. Cache size